+++ src/routes/Detail.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import DetailProductForm from '../components/Detail/DetailProductForm';
+import DetailProductInfo from '../components/product/DetailProductInfo';
+import DetailComments from '../components/comment/DetailComments';
 import styled from 'styled-components';
 
-const DetailForm = () => {
+const Detail = () => {
     return (
         <Wrap>
-            <Title>
-                <h1>상품 등록</h1>
-            </Title>
-            <DetailProductForm></DetailProductForm>
+            <Title>{/* <h1>상품 상세 페이지</h1> */}</Title>
+            <DetailProductInfo></DetailProductInfo>
+            <DetailComments></DetailComments>
         </Wrap>
     );
 };
@@ -26,8 +26,8 @@ const Wrap = styled.div`
 
 const Title = styled.div`
     /* width: 100%; */
-    margin-top: 100px;
-    margin-bottom: 50px;
+    /* margin-top: 100px; */
+    margin-bottom: 30px;
     height: 50px;
     /* background-color: red; */
     display: flex;
@@ -36,4 +36,4 @@ const Title = styled.div`
     align-items: center;
 `;
 
-export default DetailForm;
+export default Detail;
